removed unnecessary cast
git-svn-id: https://svn.code.sf.net/p/cogkit/svn/trunk/current@1990 5b74d2a0-fa0e-0410-85ed-ffba77ec0bde

+++ src/cog/modules/karajan/src/org/globus/cog/karajan/workflow/nodes/restartLog/LogEntry.java
@@ -35,7 +35,7 @@ public class LogEntry {
 
 	public static LogEntry build(VariableStack stack, FlowElement fe) throws ExecutionException {
 		LogEntry entry = new LogEntry();
-		entry.key = ThreadingContext.get(stack) + ":" + (Integer) fe.getProperty(FlowElement.UID);
+		entry.key = ThreadingContext.get(stack) + ":" + fe.getProperty(FlowElement.UID);
 		return entry;
 	}
